+++ app/controllers/application_controller.rb
@@ -14,13 +14,15 @@ class ApplicationController < ActionController::Base
   # Two things can happen when signing in.
   #   1. The current user has no email: this happens on LDAP registration. In
   #      this case, the user will be asked to submit an email.
-  #   2. Everything is fine, go to the root url.
-  def after_sign_in_path_for(_resource)
-    current_user.email? ? root_url : edit_user_registration_url
+  #   2. If the sign-in action was triggered trying to access a specific page
+  #      go back to it
+  #   3. Or go to the root url.
+  def after_sign_in_path_for(resource)
+    current_user.email? ? super : edit_user_registration_path
   end
 
   def after_sign_out_path_for(_resource)
-    new_user_session_url
+    new_user_session_path
   end
 
   def fixes
@@ -32,7 +34,7 @@ class ApplicationController < ActionController::Base
     {}.tap do |fix|
       fix[:ssl]                                = check_ssl
       fix[:secret_key_base]                    = secrets.secret_key_base == "CHANGE_ME"
-      fix[:secret_machine_fqdn]                = secrets.machine_fqdn.nil?
+      fix[:secret_machine_fqdn]                = APP_CONFIG["machine_fqdn"]["value"].blank?
       fix[:secret_encryption_private_key_path] = secrets.encryption_private_key_path.nil?
       fix[:secret_portus_password]             = secrets.portus_password.nil?
       fix
@@ -55,7 +57,7 @@ class ApplicationController < ActionController::Base
     return unless current_user && !current_user.email?
     return if protected_controllers?
 
-    redirect_to edit_user_registration_url
+    redirect_to edit_user_registration_path
   end
 
   # Redirect admin users to the registries#new page if no registry has been