+++ app/controllers/auth/registrations_controller.rb
@@ -24,7 +24,7 @@ class Auth::RegistrationsController < Devise::RegistrationsController
       sign_up(resource_name, resource)
       respond_with resource, location: after_sign_up_path_for(resource)
     else
-      redirect_to new_user_registration_url,
+      redirect_to new_user_registration_path,
         alert: resource.errors.full_messages
     end
   end
@@ -45,10 +45,10 @@ class Auth::RegistrationsController < Devise::RegistrationsController
     end
 
     if success
-      redirect_to edit_user_registration_url,
+      redirect_to edit_user_registration_path,
         notice: "Profile updated successfully!"
     else
-      redirect_to edit_user_registration_url,
+      redirect_to edit_user_registration_path,
         alert: resource.errors.full_messages
     end
   end
@@ -107,5 +107,10 @@ class Auth::RegistrationsController < Devise::RegistrationsController
   def user_params
     params.require(:user)
       .permit(:password, :password_confirmation, :current_password)
+  end
+
+  # Prevents redirect loops
+  def after_sign_up_path_for(resource)
+    signed_in_root_path(resource)
   end
 end